Update challenges page, content collections and Header component
Challenges page was updated to resolve CSS conflict affecting the header and the title section was simplified. A scheme was defined for the challenge collections and the 'showOnHomePage' field was created to indicate which challenges should be shown on the home page. The home challenges will be displayed ordered by start date.